Created a Mosaiq verification library
This library can be used to verify that a RayStation plan has been successfully exported to Mosaiq.

It compares the parameters of the the RayStation plan (beam set, beams and segments) with corresponding values found for the same patient and plan in the Mosaiq database.

Discrepancies will be listed in a GUI.