Correctly load setup offsets from prescription
It turns out we've been trying to load the setup offsets from the site setup, which is incorrect (it should be loaded from the prescription / Rad Rx object instead).

This fixes a bug where the wrong version of the setup offsets were loaded in the verification script for a plan which had been updated with a newer version in Mosaiq.