Replaced the objectives settings file
Previously we have had objectives stored in a huge settings file, which has been somewhat difficult to digest and maintain.

Now these settings has been moved to objectives classes, where separate classes exists for the various anatomical regions. This is a more object oriented approach which should make it easier to maintain and to furher simplify the code in the future,