More robust handling of SOM group creation
In the initial version, a simulate organ motion group would be created whenever the Plan script was executed. This would obviously be wrong in the case the plan script is run multiple times. Now it will check if a SOM group exists for the given examination, and only create a new SOM group if one with the same examination reference doesnt already exist.