Add dialogue for breast robustness
Previously in the PLAN script, we implemented an infrerior robustness margin for breast when breast size was larger than 850 cm^3. Such a hard limit turned out to be imprecise in practice.

With this update, we replace this size based limit with a GUI dialogue, asking the user for his preference on whether to use inferior robustness for the given case.