Faster DLS ROI creation
It turns out our implementation of creating DLS ROIs one by one was extremely slow.

The code has now been refactored to collect the DLS ROIs in a group and create them all in once, which increases the speed considerably.

Example benchmark:
Time consumption for a prostate case with lymph nodes was reduced from 9 to 2 minutes.